+++ app/src/utils/moveBlock/moveBlock.js
@@ -265,3 +265,20 @@ const alreadyRequested = (move, lastMoveRequest) => {
     return false;
   return true;
 };
+
+/** Get the lock of the player with a given id
+ *
+ * @param {*} players : list of all players with lock status
+ * @param {*} pid : id of peer
+ * @returns
+ */
+export const getLock = (players, pid) => {
+  for (let p of players) {
+    if (!p.hasOwnProperty('lock')) {
+      return false;
+    }
+    if (p.id === pid) {
+      return p.lock;
+    }
+  }
+};
